backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
langchain==0.2.16
langchain-openai==0.1.24
langchain-core==0.2.39
langchain-community==0.2.16
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==3.2.2
python-dotenv==1.0.0
openai>=1.26.0
pandas==2.0.3